[HER-104] Change score type to float (#67)
* [HER-104] Change score type to float in entity

* [HER-104] Change score type in models

* [HER-104] Change score type in PDetection

* [HER-104] Create migration for changing score type

* [HER-104] Add some floating numbers in detections

* [HER-104] Fix loss of precision in comparison of floating numbers

* [HER-104] Follow-up Change formula

+++ src/Business/Detection/Creating/Models/MCreateDetection.cs
@@ -4,7 +4,7 @@ namespace Business.Detection.Creating.Models
 
     public class MCreateDetection
     {
-        public int Score { get; set; }
+        public float Score { get; set; }
 
         public string Class { get; set; }
